# Schematics performance