--- a/modules/effects/schematics/ng-add/index.ts
+++ b/modules/effects/schematics/ng-add/index.ts
@@ -137,15 +137,7 @@ export default function(options: RootEffectOptions): Rule {
 
     return chain([
       branchAndMerge(
-        chain([
-          filter(
-            path =>
-              path.endsWith('.module.ts') &&
-              !path.endsWith('-routing.module.ts')
-          ),
-          addImportToNgModule(options),
-          mergeWith(templateSource),
-        ])
+        chain([addImportToNgModule(options), mergeWith(templateSource)])
       ),
       options && options.skipPackageJson
         ? noop()