--- a/modules/schematics/src/store/index.ts
+++ b/modules/schematics/src/store/index.ts
@@ -6,7 +6,6 @@ import {
   apply,
   branchAndMerge,
   chain,
-  filter,
   mergeWith,
   template,
   url,
@@ -166,15 +165,7 @@ export default function(options: StoreOptions): Rule {
 
     return chain([
       branchAndMerge(
-        chain([
-          filter(
-            path =>
-              path.endsWith('.module.ts') &&
-              !path.endsWith('-routing.module.ts')
-          ),
-          addImportToNgModule(options),
-          mergeWith(templateSource),
-        ])
+        chain([addImportToNgModule(options), mergeWith(templateSource)])
       ),
     ])(host, context);
   };