Closes #1102

Some stats on an empty project:
Effects: before 8s, with these changes 2s
Store: before 6s, with these changes 1s
Reducer: before 9s, with these changes 1s
Feature: before 15s, with these changes 3s